requirements.txt
# Required Python packages for the Grape Labeling Tool

# Core
ultralytics>=8.0.0  # For YOLOv8
opencv-python>=4.5.0
Pillow>=8.0.0

# GUI
tk>=0.1.0  # Usually included with Python, but listed for clarity

# Data Processing
numpy>=1.20.0
pandas>=1.3.0

# Optional - If using regression model in the tool
scikit-learn>=1.0.0
